perf: :adhesive_bandage: Improve the keyframe selection algorithm
In this commit, I fixed the code for the smart Keyframe detection algorithm. Basically, now it maintains that the distance between any two selected keyframes is not less than or greater than defined criteria. I still bucket the frames, select the peaks from the buckets with a threshold of 0, and minimum frmaes apart. This is all done by peakutils library itself. Then, I do post processing on selected frames to make sure that if two frmaes are greater than a defined criteria apart, I inject frames in between to to reduce this while still making sure that the limit of minimum frames in between is kept in check. I also make sure to always have the first and last frame in keyframes.